Make a test in ClientTest more consistent
ClientTest->testDeleteIdsIdxStringTypeString assumed that new documents
created without a specified id or routing value would route to the same
shard as "1" and a different shard from "2" when there are two shards.
This wasn't always that case because:
1.  Documents without a routing value use their ids as a route.
2.  Documents that don't specify an id get a random one.

This sets the routing for the document to "1" which, luckily, does in
fact route to a different shard then "2" when there are two shards.  And
it will stay that way forever back Elasticsearch has to maintain a consistent
routing style or it won't be safe to upgrade.

Closes #633