Avoid remove previously added params where add a suggest to the query
When we add a suggest to the query (while wanting to keep the query part
of the query), the query and all the previoulsy added parameters where
removed.

Like in the unit test, if we add the query (or the size) then the suggest,
the query (or the size) is removed. But if we add the suggest then the query
(or the size), the query is still present.

So insted of removing all the params, we merge the suggest with the
existing parameters.