Test case for script-as-string
My change slightly dropped code coverage: when passing a Script, it used to
manually extract the script (as string) from the Script object & set it.
Now that I've changed it to use Script::toArray(), that "set script as string"
patch was no longer being tested.
This adds a separate test for that case.