Add tests for Bulk hasIndex() and hasType()
The two functions wern't returning what i would expect, because
the tests check explicitly for the empty string and the default values
were null. Added a couple tests to assert their proper behaviour and
then fixed the code to pass the tests.

Additionally adjusts the type annotations on index and type to reflect
their ability to return null values. This looks to be consistent with
other parts of the codebase that also annotate things with string|null.