First step to upgrade Elastic to elasticsearch 5.0 (#1114)
In this first step, all tests are checked which are not compatible with elasticsearch 5.0. Either the tests are directly fixed or they are temporarely skipped and have to be cleaned up later. In general it seems as most of the things will just keep working with elasticsearch 5.0 except for the things that were changed / removed in elasticsearch 5.0 anyways (like mapping changes from string to text / keyword).

* Replace remove count search type by query_then_fetch
* Replace optimize by forcemerge as optimize was removed
* Rename getMock() to createMock() to be compatible with newer phpunit versions
* Convert string mapping to text
* Change delete-by-query implementation to use 5.0 api
* Remove delete-by-query plugin checks as now installed by default
* Change script implementation to new 5.0 implementation
* Simplify update-document
* Update test environment to use self built elasticsearch instance
* Improve node info fetching to use new format
* Skip plugin tests if no plugin available
* Fix list of tests
* Apply liniting
* Fix script tests for new script structure
* Remove search_type from scroll as not needed anymore
* Update README.md with new dependency

Part of https://github.com/ruflin/Elastica/issues/1184