Remove support for PHP 5.4 and PHP 5.5
Elastica now requires at least PHP 5.6. The reasons for the removal is that PHP 5.4 and 5.5 are EOL. 5.6 will allow Elastica to use new features and also update its dependencies to the most recent versions.

Elastica probably still works with PHP 5.4 and PHP 5.5 but is not tested anymore and composer.json will require 5.6.