Properly pass query options with Multi\Search (#1224)
According to the doc only few options need to be passed in the header of a
MultiSearch. Unfortunately the current behavior is to sent all query options
into the header. For instance terminate_after or timeout are ignored by elastic
if the search is sent via a MultiSearch.
This patch adds an explicit filter of query options to pass in the header and
write remaining ones in the search request body.