Finalize elasticsearch 5 migration (#1242)
* Remove outdated features that do not exist anymore in elasticsearch 5
* geo_distance_range and geohash_cell are deprecated in ES 5.0 and only work with indexes created in 2.x
* deprecate Match::setFieldType
* fix bool query test and dsl tests
* fix match and indices tests
* fix dynamic template test
* Fix multi search test with partial error

The previous bad query now throws a syntax error from ES due to an invalid range option which makes the whole multi search fail instead of only a single search within the multi search. So I replaced it with a different kind of error that only fails for this single search

* fix aggregation tests
* fix InnerHit tests

see https://github.com/elastic/elasticsearch/issues/21312

* fix MoreLikeThis tests and remove leftover filter logic
* fix geo distance tests and inner hits cannot be fixed due to ES bug
* There is not routing parameter when creating an index (and has never been).

So this feature has always been useless but ES 5 now validates the wrong param. Creating aliases has the possibility to specify routing and a filter. But the alastica api doesn't provide those params yet, which could be added in the future

* fix Bulk tests
* fixing some script logic and introducing ScriptId

ScriptId is used to reference stored scripts ("id" property has been renamed to "stored" in ES 5.1 which we need to do as well in later releases)

* cleanup duplicated tests for old deprecated ES test classes

and deprecate null transport for transition to NullTransport just as Bool is deprecated in favor of BoolQuery

* add changelog entries