Type.php: Removed Exception thrown in deleteById()
An exception was thrown if no data set corresponding to param $id was found (and hence was not deleted). Removed this behavior because it is a very harsh reaction to a more or less uncritical situation because the result is the same: The data requested to be deleted no longer exists.