Implement Pipeline and Processors (#1372)
This has been a long implementation, Pipelines need to be implemented and even processor if we want to use field attachments.

- Implementation of Pipeline
- Implementation of Processors (only 4 processors are missing : **Foreach, Grok, Gsub, Script**)

This is just a first implementation : I have some doubt on extending **Param** for Processors and for Pipeline Object. It's very useful, 'cause extending ArrayableInterface Object are quickly converted to Arrays.