Implement an ErrorTransport (#1529)
When using AwsAuthV4 transport, it is possible that the user provides incorrect credentials. 

This means that the Guzzle client will get a 403 from AWS. 

By default Guzzle silently ignores all exceptions and stores them in `$response->setTransferInfo()`

This feature allows a downstream client/library to handle them however. 

Currently, FOSElastica does nothing to process 403 or other errors that come back from ruflin/Elastica (transferInfo is never used). 

This causes a problem because 403 are silently ignored and the progress bar appears to be moving unhindered ;-) 

This PR is a pre-requisite for unit tests that will accompany this fix on FOSElastica