Make classes with factory methods extensible (#1662)
# What this is changing
Use `static` instead of `self` in all factory methods in Elastica.

# What is my motivation/problem
I am working on a ES persistence layer library for my company.
My goal is to write it in a way that users can combine various query implementations with various ES clients (i.e. elasticsearch-php and elastica at the moment), Elastica queries being among the options. For this i wrote a thin wrapper for the `Elastica\Query` class and struggled with the factory method since it was using `self` instead of `static`. This is violating the open-closed principle in my opinion, hence i created this PR :)