Added Collapse DSL to QueryBuilder (#1724)
That's going to be a bit awkward now: in the PR that added Field Collapsing Support (#1653) I forgot to check in the DSL classes, which made working with it a bit clumsy compared to the usual query methods.
I.e. - as with aggregations - it's now possible to do stuff like this:
```
$collapse
  ->setFieldname('product_id')
  ->setInnerHits(
    $qb->collapse->inner_hits()  // instead of (new Collapse\InnerHits())
  )
```

Let me know if there need to be any changes. Cheers :)